requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blis==1.3.0
boto3==1.40.63
botocore==1.40.63
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
deeplake==3.9.52
dill==0.4.0
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.20.0
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.36.0
humbug==0.3.2
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.0.2
langchain-openai==1.0.1
langcodes==3.5.0
langsmith==0.4.38
language_data==1.3.0
lz4==4.4.4
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multiprocess==0.70.18
murmurhash==1.0.13
nltk==3.9.2
numpy==2.3.4
openai==2.6.1
orjson==3.11.4
packaging==25.0
pathos==0.3.4
pillow==10.4.0
pox==0.3.6
ppft==1.7.7
preshed==3.0.10
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
smart_open==7.4.1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==2.0.0
zstandard==0.25.0
